Addressed feedback - Renamed TagBindingResult to TagHelperBinding - Added reference to TagHelperBinding in SetTagHelperPropertyIRNode - Cleanup
