+++ src/Microsoft.AspNetCore.Razor.Evolution/Intermediate/SetPreallocatedTagHelperPropertyIRNode.cs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Collections.Generic;
+using Microsoft.AspNetCore.Razor.Evolution.Legacy;
 
 namespace Microsoft.AspNetCore.Razor.Evolution.Intermediate
 {
@@ -22,6 +23,8 @@ namespace Microsoft.AspNetCore.Razor.Evolution.Intermediate
         public string PropertyName { get; set; }
 
         public BoundAttributeDescriptor Descriptor { get; set; }
+
+        internal TagHelperBinding TagHelperBinding { get; set; }
 
         public bool IsIndexerNameMatch { get; set; }
