+++ src/Microsoft.AspNetCore.Razor.Evolution/Intermediate/SetTagHelperPropertyIRNode.cs
@@ -25,6 +25,8 @@ namespace Microsoft.AspNetCore.Razor.Evolution.Intermediate
 
         public BoundAttributeDescriptor Descriptor { get; set; }
 
+        internal TagHelperBinding TagHelperBinding { get; set; }
+
         public bool IsIndexerNameMatch { get; set; }
 
         public override void Accept(RazorIRNodeVisitor visitor)
