+++ src/Microsoft.AspNetCore.Razor.Evolution/RazorPreallocatedTagHelperAttributeOptimizationPass.cs
@@ -139,6 +139,7 @@ namespace Microsoft.AspNetCore.Razor.Evolution
                     TagHelperTypeName = node.TagHelperTypeName,
                     PropertyName = node.PropertyName,
                     Descriptor = node.Descriptor,
+                    TagHelperBinding = node.TagHelperBinding,
                     Parent = node.Parent,
                     IsIndexerNameMatch = node.IsIndexerNameMatch,
                 };
